VSTR-134 Test InfoMockTest 설정 분리
